Create SaleProject, Details and Confirmation function component and move related code blocks respectively.
